Proof read
It might be nice to have a little flow chart of what is happening at the various clicks when using git (pull, stage, commit, pull, push), maybe not necessary now but would be nice. 

A screenshot highlighting the pull, push, bit on line 82.

A screenshot highlighting the History click on line 88.