Update 2017-03-15-mixed-models.md
You might want to rethink your section on model selection: "EXTRA: P-values and model selection" - "THE END" it seems like there is some jumping around and digression that could be fixed.

I assumed you are talking about Zuur et al. 2009 "Mixed Effects Models and Extensions in Ecology with R". Correct if I'm wrong. Is there a link to a publicly available pdf of this info?