Update 2018-01-25-qualitative.md
Hey John,

Thanks for the tutorial! Just a few quick thoughts - can you please add a survey link?

I have edited the text slightly - I removed the AIC bit at the end, just because I think if we are to delve into model comparison, there should be more details about what it is and when it's appropriate to use it. I also changed some of the phrasing, mostly where there was "unfortunately...", just because that's how things are, no need to be bummed out about it.